Separate webhook tests file
